Backend implemented using Minimal API

+++ APIDotNetCore/DataLayer/Models/Entities/Tasks.cs
@@ -11,7 +11,7 @@ namespace DataLayer.Models.Entities
         public long id { get; set; }
         public string? title { get; set; }
         public string? details { get; set; }
-        public int? progress_ratio { get; set; }
+        public double? progress_ratio { get; set; }
         public DateTime? insert_date { get; set; }
     }
 }
